feature/#184: move configurable routes under /routing folder in core
