+++ projects/core/public_api.ts
@@ -6,5 +6,4 @@ export * from './src/config/index';
 export * from './src/state/index';
 export * from './src/site-context/index';
 export * from './src/occ/index';
-export * from './src/route/index';
 export * from './src/routing/index';
