+++ projects/core/src/routing/configurable-routes/configurable-route.ts
