+++ projects/core/src/routing/configurable-routes/configurable-routes-module.config.ts
