+++ projects/core/src/routing/configurable-routes/index.ts
@@ -2,5 +2,5 @@ export * from './configurable-routes.service';
 export * from './configurable-routes.module';
 export * from './configurable-routes-module.config';
 export * from './configurable-routes.service';
-export * from './configurable-routes';
+export * from './configurable-route';
 export * from './routes-config';
