+++ projects/core/src/routing/configurable-routes/routes-config.ts
