+++ projects/core/src/routing/index.ts
@@ -3,3 +3,5 @@ export * from './config/routing-module-config';
 export * from './facade/routing.service';
 
 export * from './models/page-context.model';
+
+export * from './configurable-routes/index';
