+++ projects/core/src/routing/routing.module.ts
@@ -25,7 +25,7 @@ import { RoutingService } from './facade/routing.service';
 
 import { ROUTING_FEATURE } from './state';
 import { defaultConfig } from './config/default-config';
-import { ConfigurableRoutesModule } from '../route/configurable-routes.module';
+import { ConfigurableRoutesModule } from './configurable-routes/configurable-routes.module';
 
 export function getMetaReducers(
   config: RoutingModuleConfig
